backend/pyproject.toml
[project]
name = "rag-chatbot-backend"
version = "0.1.0"
description = "RAG Chatbot Backend for Physical AI Textbook"
requires-python = ">=3.11"
dependencies = [
    "asyncpg>=0.31.0",
    "fastapi>=0.124.0",
    "google-generativeai>=0.8.5",
    "greenlet>=3.3.0",
    "openai>=2.9.0",
    "openai-agents[sqlalchemy]>=0.6.2",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "qdrant-client>=1.16.1",
    "sqlalchemy>=2.0.44",
    "uvicorn[standard]>=0.38.0",
]

[dependency-groups]
dev = [
    "black>=25.11.0",
    "httpx>=0.28.1",
    "mypy>=1.19.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.14.8",
]
